Suggested API changes
