+++ src/Compilers/Core/Portable/CodeAnalysis.csproj
@@ -51,7 +51,7 @@
     <Compile Include="Binding\AbstractLookupSymbolsInfo.cs" />
     <Compile Include="CaseInsensitiveComparison.cs" />
     <Compile Include="CodeGen\ILEmitStyle.cs" />
-    <Compile Include="Compilation\CompilationStuff.cs" />
+    <Compile Include="Compilation\CompilationContext.cs" />
     <Compile Include="DiagnosticAnalyzer\AnalyzerDriver.CompilationData.cs" />
     <Compile Include="DiagnosticAnalyzer\SuppressMessageInfo.cs" />
     <Compile Include="Diagnostic\SuppressionInfo.cs" />
