+++ src/Compilers/Core/Portable/Compilation/Compilation.cs
@@ -33,9 +33,9 @@ namespace Microsoft.CodeAnalysis
     public abstract partial class Compilation
     {
         /// <summary>
-        /// For use only in <see cref="Microsoft.CodeAnalysis.Collections.CompilationStuff"/>.
+        /// For use only in <see cref="Microsoft.CodeAnalysis.Collections.CompilationContext"/>.
         /// </summary>
-        internal ConcurrentDictionary<object, object> _compilationStuff = new ConcurrentDictionary<object, object>();
+        internal ConcurrentDictionary<object, object> _compilationContext = new ConcurrentDictionary<object, object>();
 
         /// <summary>
         /// Returns true if this is a case sensitive compilation, false otherwise.  Case sensitivity
